moving channel nomenclature to tf
